Dmenu can use dots for marked items and can pre mark items
Any arbitary character can be used to denote the marked items. And items
starting with "***" followed by a tab character is marked as a marked
item if pre_mark is enabled. Also added various flags are added to
control each funtionality such as use_dots, pre_mark... But using dots
might overwrite the text if height is too low.