Cleaned up code which draws dmenu grid
It might show a 1 pixel gap at the last column, and definetly we need to
test it how it works. This one probablily fixed the gap which was showing
at the end of last column and the border caused by the turnication of int
typed varible. Previously we were simply filling up the last column. But
after this one we should have columns with equal width.